pyproject.toml
[project]
name = "agenti-bluuu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "google-genai>=1.67.0",
    "langchain>=1.2.12",
    "langchain-google-genai>=2.1.12",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.2",
    "langsmith>=0.7.17",
    "opencv-python>=4.13.0.92",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.2.2",
    "ultralytics>=8.4.21",
    "uvicorn>=0.41.0",
]
